% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtransf.R
\name{devtransf}
\alias{devtransf}
\title{calculates derived variables by deviation transformation}
\usage{
devtransf(ev, optimum, devexp = c(0.5, 1, 2))
}
\arguments{
\item{ev}{Vector of values to be transformed}

\item{optimum}{Center value from which deviation is calculated}

\item{devexp}{Vector of exponent values to scale the steepness of the
deviation transformation}
}
\description{
Note that this function calculates devation AND rescales to [0,1]
}

