% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvfromev.R
\name{dvfromev}
\alias{dvfromev}
\title{produces dvs from a given ev by different transformations.}
\usage{
dvfromev(df, writedir, transformtype, allsplines)
}
\arguments{
\item{df}{Dataframe with 2 columns: response variable and explanatory
variable (in that order). Column names are used as identifiers. The
response variable represents presence or background, coded as: 1/NA. The
explanatory variable may be continuous or categorical.}

\item{writedir}{Directory to which Maxent runs of spline transformations are
written}

\item{transformtype}{Set of transformation types to be used.}

\item{allsplines}{Logical. Keep all spline transformations.}
}
\value{
Dataframe with one column for each DV.
}
\description{
In MIAT, "D" transformation is only performed if the optimum occurs in the
middle 80\% of the EV range. \code{dvfromev} does not specify this condition.
}

