% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DVs}
\alias{DVs}
\title{Derived variables and transformation functions.}
\format{List with 2 elements: \enumerate{ \item A list of 4 data frames, with
  each containing the derived variables produced for a given explanatory
  variable. \item A list of all the transformation functions used to produce
  the derived variables. }}
\source{
produced from \code{\link{sp1po}} using \code{\link{deriveVars}}
}
\usage{
DVs
}
\description{
Derived variables and transformation functions for distribution modeling of a
simulated data set.
}
\keyword{datasets}

