% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewma.R
\name{ewma}
\alias{ewma}
\title{calculates exponentially weighted moving average}
\usage{
ewma(x, n)
}
\arguments{
\item{x}{numeric. Vector across which the moving average is to be applied.}

\item{n}{integer. Width of the moving average window. Should be odd,
otherwise the window will be uncentered.

Imports: stats}
}
\description{
calculates exponentially weighted moving average
}

