% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fopoptimum.R
\name{fopoptimum}
\alias{fopoptimum}
\title{calculate optimum ev value based on frequency of presence}
\usage{
fopoptimum(data, intervals = 20, smoothwindow = 5, EVranging = F)
}
\arguments{
\item{data}{Dataframe containing the response variable in the first column and
explanatory variables in the second column. The response variable should
represent presence or background, coded as: 1/NA.}

\item{intervals}{Number of intervals into which the continuous EV is divided.
Irrelevant for categorical EVs.}

\item{smoothwindow}{Width of the smoothing window (in an exponentially
weighted moving average). Irrelevant for categorical EVs.}

\item{EVranging}{if \code{TRUE}, will range the EV scale to [0,1]. Irrelevant
for categorical EVs.}
}
\value{
the EV value at which FOP is highest (\code{EVoptimum})
}
\description{
The optimum that is retuned is based on the smoothed data, unless a maximum
exists at the extremes of the EV (outside the 5-interval smoothing window).
}
\details{
\code{DESCRIPTION Imports}: dplyr, Hmisc, scales
}

