% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grasslandDVs}
\alias{grasslandDVs}
\title{Derived variables and transformation functions.}
\format{List with 2 elements: \enumerate{ \item A list of 13 data frames,
  with each containing the derived variables produced for a given explanatory
  variable. \item A list of all the transformation functions used to produce
  the derived variables. }}
\source{
produced from \code{\link{grasslandPO}} using
  \code{\link{deriveVars}}
}
\usage{
grasslandDVs
}
\description{
Derived variables and transformation functions for modeling the distribution
of semi-natural grasslands in Oestfold County, Norway.
}
\keyword{datasets}

