% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minskew.R
\name{minskew}
\alias{minskew}
\title{calculates skewness of a vector}
\usage{
minskew(x)
}
\arguments{
\item{x}{Vector of data. Must have scale [0,1]!}
}
\description{
Also calculates the constant 'c' needed for zero-skewness transformation in
\code{scalex}
}
\details{
\code{DESCRIPTION Imports}: e1071
}

