% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFOP.R
\name{plotFOP}
\alias{plotFOP}
\title{Plot Frequency of Observed Presence (FOP).}
\usage{
plotFOP(data, EV, span = 0.5, intervals = NULL, ranging = FALSE)
}
\arguments{
\item{data}{Data frame containing the response variable in the first column
and explanatory variables in subsequent columns. The response variable
should represent presence or background, coded as: 1/NA. See Details for
information regarding presence/absence data. See also
\code{\link{readData}}.}

\item{EV}{Name or column index of the explanatory variable in \code{data} for
which to calculate FOP.}

\item{span}{The proportion of FOP points included in the local regression
neighborhood. Should be between 0 and 1. Irrelevant for categorical EVs.}

\item{intervals}{Number of intervals into which the continuous EV is divided.
Defaults to the minimum of N/10 and 100. Irrelevant for categorical EVs.}

\item{ranging}{Logical. If \code{TRUE}, will range the EV scale to [0,1].
This is equivalent to plotting FOP over the linear transformation produced
by deriveVars. Irrelevant for categorical EVs.}
}
\value{
In addition to the graphical output, a list of 2: \enumerate{ \item
  \code{EVoptimum}. The EV value (or level, for categorical EVs) at which FOP
  is highest \item  \code{FOPdata}. A data frame containing the plotted data.
  Columns in this data frame represent the following: EV interval ("int"),
  number of points in the interval ("n"), mean EV value of the points in the
  interval ("intEV"), mean RV value of the points in the interval ("intRV"),
  and exponentially weighted moving average of intRV ("smoothRV"). For
  categorical variables, only the number of points in the level ("n"), the
  level name ("level"), and the mean RV value of the level ("levelRV") are
  used.}
}
\description{
\code{plotFOP} produces a Frequency of Observed Presence (FOP) plot for a
given explanatory variable. An FOP plot shows the rate of occurrence of the
response variable across intervals or levels of the explanatory variable. For
continuous variables, a local regression ("loess") of the FOP values is added
to the plot as a line. \code{plotFOP} also returns a list containing the
optimum EV value, and a data frame containing the plotted data (for
customizable plotting).
}
\details{
In the local regression ("loess"), the plotted FOP values are regressed
against the the EV values. The points are weighted by the number of
observations they represent, such that an FOP value in an interval with many
observations is given more weight.

For continuous variables, the returned value of 'EVoptimum' is based on the
loess-smoothed FOP values, such that a point maximum in FOP may not always be
considered the optimal value of EV.

If the response variable in \code{data} represents presence/absence data, the
result is an empirical frequency of presence curve, rather than a observed
frequency of presence curve (see Stoea et al. [in press], Sommerfeltia).
}
\examples{

}
\references{
Stoea, B., Halvorsen, R., Mazzoni, S. & Gusarov, V. (2016)
  Sampling bias in presence-only data used for species distribution
  modelling: assessment and effects on models. Sommerfeltia, submitted
  manuscript.
}
