% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResp.R
\name{plotResp}
\alias{plotResp}
\title{Plot model response across a given explanatory variable.}
\usage{
plotResp(data, EV, dvdata, dir = NULL, jarpath = NULL, logscale = FALSE,
  ...)
}
\arguments{
\item{data}{Data frame of training data, with response variable (1/NA) in the
first column and explanatory variables in subsequenct columns.}

\item{EV}{Name or list index of the explanatory variable in \code{dvdata} for
which the response curve is to be generated. Interaction terms not allowed.}

\item{dvdata}{List of derived variables used to train the model, with each
list item a data frame containing 1 or more DVs for a given EV. E.g. output
[[1]] of \code{selectEV}.}

\item{dir}{Directory to which Maxent files will be written. Defaults to
the working directory.}

\item{jarpath}{Pathway to the maxent.jar executable jar file. If
unspecified, the function looks for the file in \code{dir}.}

\item{logscale}{logical. Plot the common logarithm of PRO rather than PRO.}

\item{...}{Arguments to be passed to \code{plot} to control the appearance of
the points in the scatterplot. For example: \itemize{ \item \code{cex} for
size \item \code{col} for color \item \code{pch} for type }}
}
\value{
In addition to the graphical output, the plotted data is returned.
}
\description{
\code{plotResp} plots  the response of a given Maxent model over any of the
included explanatory variables (EVs) in that model. For categorical
variables, a box plot is returned rather than a curve. \code{plotResp} also
returns a data frame containing the plotted data (for customizable graphics).
}
\details{
The response curves generated in this function are single-effect response
curves, presenting the response of a model containing the explanatory
variable of interest only (cf. marginal-effect response curves).

The \code{ev} specified in \code{dvdata} must not be an interaction term.
}

