% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResp2.R
\name{plotResp2}
\alias{plotResp2}
\title{Plot marginal-effect model response.}
\usage{
plotResp2(data, EV, transformation, model, logscale = FALSE, ...)
}
\arguments{
\item{data}{Data frame containing the response variable in the first column
and explanatory variables in subsequent columns. The response variable
should represent presence/background data, coded as: 1/NA.}

\item{EV}{Name or column index of the explanatory variable in \code{data} for
which the response curve is to be generated.}

\item{transformation}{Full pathway of the 'transformations.Rdata' file
containing the transformations used to build the model. This file is saved
as a result of the \code{\link{deriveVars}} function. Equivalently, the
second item in the list returned by \code{\link{deriveVars}} can be used
directly.}

\item{model}{Full pathway of the '.lambdas' file of the model in question.
This file is saved as a result of \code{\link{selectEV}}.}

\item{logscale}{Logical. Plot the common logarithm of PRO rather than PRO
itself.}

\item{...}{Arguments to be passed to \code{plot} to control the appearance of
the plot. For example: \itemize{ \item \code{cex} for size of points \item
\code{col} for color \item \code{pch} for type }}
}
\value{
In addition to the graphical output, the plotted data is returned. In
  the case of a continuous EV, the plotted data consists of both individual
  points ('respPts') and the smoothed moving average of those points
  ('respLine').
}
\description{
\code{plotResp2} plots the marginal-effect response of a given Maxent model
over any of the included explanatory variables (EVs) in that model. For
categorical variables, a bar plot is returned rather than a scatter plot.
\code{plotResp2} also returns a data frame containing the plotted data (for
customizable graphics). Marginal-effect response curves present the response
of the model when all other explanatory variables are held constant at their
mean values (cf. single-effect response curves; \code{\link{plotResp}}).
}

