% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runIWLR.R
\name{predict.iwlr}
\alias{predict.iwlr}
\title{S3 method for class "iwlr": returns model predictions for new data in "PRO"
or "raw" format.}
\usage{
\method{predict}{iwlr}(model, newdata, type = "PRO")
}
\arguments{
\item{model}{Model of class "iwlr"}

\item{newdata}{Data frame containing variables to predict across}

\item{type}{Type of model output: "PRO" or "raw"}
}
\description{
S3 method for class "iwlr": returns model predictions for new data in "PRO"
or "raw" format.
}
