% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runIWLR.R
\name{predict.MIAmaxent_iwlr}
\alias{predict.MIAmaxent_iwlr}
\title{Predict method for infinitely-weighted logistic regression}
\usage{
\method{predict}{MIAmaxent_iwlr}(object, newdata, type = "PRO", ...)
}
\arguments{
\item{object}{Model of class "MIAmaxent_iwlr"}

\item{newdata}{Data frame containing variables with which to predict}

\item{type}{Type of model output: "PRO" or "raw"}
}
\description{
Returns model predictions for new data in "PRO" or "raw" format.
}
\keyword{internal}
