% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seldvs}
\alias{seldvs}
\title{Selected derived variables accompanied by selection trails.}
\format{List with 2 elements: \enumerate{ \item A list of 3 data frames, with
  each containing the derived variables selected for a given explanatory
  variable. \item A list of forward model selection trails used to select
  derived variables. }}
\source{
produced from \code{\link{DVs}} and \code{\link{sp1po}} using
  \code{\link{selectDVforEV}}
}
\usage{
seldvs
}
\description{
Selected derived variables and tables showing forward model selection of
derived variables for distribution modeling of a simulated data set.
}
\keyword{datasets}

