% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectDVforEV.R
\name{selectDVforEV}
\alias{selectDVforEV}
\title{Select parsimonious sets of derived variables.}
\usage{
selectDVforEV(rv, dv, alpha = 0.01, writedir = NULL, jarpath = NULL)
}
\arguments{
\item{rv}{Response variable vector. The RV should represent
presence/background data, coded as: 1/NA.}

\item{dv}{List of data frames, with each data frame containing DVs for a
given EV. E.g. output of \code{deriveVars}.}

\item{alpha}{Alpha-level used in F-test comparison of models. Default is
0.01.}

\item{writedir}{The directory to which Maxent files will be written during
subset selection of DVs. Defaults to the working directory.}

\item{jarpath}{The pathway to the maxent.jar executable jar file. If
unspecified, the function looks for maxent.jar in the writedir.}
}
\value{
List of length two. The first item is a list of data frames, with
  each data frame containing \emph{selected} DVs for a given EV. The second
  item is also a list of data frames, where each data frame shows the trail
  of forward selection for a given EV.
}
\description{
For each explanatory variable (EV), \code{selectDVforEV} selects the parsimonious
set of derived variables (DV) which best explains variation in a given
response variable. The function uses a process of forward selection based on
comparison of nested models by the F-test, where the F-statistic is
calculated using equation 59 in Halvorsen (2013). See Halvorsen et al. (2015)
for an explanation of the forward selection procedure.
}
\details{
If the derived variables were created using \code{\link{deriveVars}}, the
same response variable should be used in \code{selectDVforEV}, as the deviation
and spline transformations produced by \code{deriveVars} are RV-specific.

DVs must be uniquely named, and the names must not contain spaces.
}
\references{
Halvorsen, R. (2013). A strict maximum likelihood explanation of
  MaxEnt, and some implications for distribution modelling. Sommerfeltia, 36,
  1-132.

Halvorsen, R., Mazzoni, S., Bryn, A., & Bakkestuen, V. (2015).
  Opportunities for improved distribution modelling practice via a strict
  maximum likelihood interpretation of MaxEnt. Ecography, 38(2), 172-183.
}

