% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{selevs}
\alias{selevs}
\title{Selected explanatory variables accompanied by selection trails.}
\format{List with 2 elements: \enumerate{ \item A list of 5 data frames,
  where the first 3 represent selected explanatory variables, and the last 2
  represent selected interaction terms between these explanatory variables.
  \item A trail of forward model selection used to select explanatory
  variables and interaction terms. }}
\source{
produced from \code{\link{seldvs}} and
  \code{\link{sp1po}} using \code{\link{selectEV}}
}
\usage{
selevs
}
\description{
Selected explanatory variables and tables showing forward model selection of
explanatory variables for distribution modeling of a simulated data set. Each
individual explanatory variable is represented by a group of derived
variables.
}
\keyword{datasets}

