% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp1po}
\alias{sp1po}
\title{Simulated occurrence and environmental data.}
\format{A data frame with 40 rows and 5 variables: \describe{
  \item{RV}{response variable, occurrence either presence or uninformed
  background} \item{EV11}{explanatory variable: northing}
  \item{EV12}{explanatory variable: easting} \item{EV13}{explanatory
  variable: modified random uniform} \item{EV14}{explanatory variable: random
  uniform} }}
\source{
Halvorsen, R. (2013) A strict maximum likelihood explanation of
  MaxEnt, and some implications for distribution modelling. Sommerfeltia, 36,
  1-132.
}
\usage{
sp1po
}
\description{
Simulated data set for distribution modeling consisting of occurrence and
environmental data. The simulated study area consists of 40 grid cells, with
8 row and 5 columns, in which 10 presences occur.
}
\keyword{datasets}

