% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toydata_dvs}
\alias{toydata_dvs}
\title{Derived variables and transformation functions, from toy data.}
\format{List with 2 elements: \enumerate{ \item A list of 5, with the response
  variable followed by data frames each containing the derived variables
  produced for a given explanatory variable. \item A list of the response
  variable and all the transformation functions used to produce the derived
  variables. }}
\source{
Produced from \code{\link{toydata_sp1po}} using
  \code{\link{deriveVars}}.
}
\usage{
toydata_dvs
}
\description{
Derived variables and transformation functions for distribution modeling of a
small, synthetic data set used in Halvorsen (2013).
}
\references{
Halvorsen, R. (2013) A strict maximum likelihood explanation of
  MaxEnt, and some implications for distribution modelling. Sommerfeltia, 36,
  1-132.
}
\keyword{datasets}
