% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toydata_seldvs}
\alias{toydata_seldvs}
\title{Selected derived variables accompanied by selection trails, from toy data.}
\format{
List with 2 elements: \enumerate{ \item A list of 3, with the
  response variable followed by data frames each containing the derived
  variables selected for a given explanatory variable. \item A list of the
  response variable and forward model selection trails used to select derived
  variables. }
}
\source{
Produced from \code{\link{toydata_dvs}} using
  \code{\link{selectDVforEV}}.
}
\usage{
toydata_seldvs
}
\description{
Selected derived variables and tables showing forward model selection of
derived variables for distribution modeling of a small, synthetic data set
used in Halvorsen (2013).
}
\references{
Halvorsen, R. (2013) A strict maximum likelihood explanation of
  MaxEnt, and some implications for distribution modelling. Sommerfeltia, 36,
  1-132.
}
\keyword{datasets}
