% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toydata_sp1po}
\alias{toydata_sp1po}
\title{Occurrence and environmental toy data.}
\format{A data frame with 40 rows and 5 variables: \describe{
  \item{RV}{response variable, occurrence either presence or uninformed
  background} \item{EV11}{explanatory variable: northing}
  \item{EV12}{explanatory variable: easting} \item{EV13}{explanatory
  variable: modified random uniform} \item{EV14}{explanatory variable: random
  uniform} }}
\source{
Halvorsen, R. (2013) A strict maximum likelihood explanation of
  MaxEnt, and some implications for distribution modelling. Sommerfeltia, 36,
  1-132.
}
\usage{
toydata_sp1po
}
\description{
A small, synthetic data set for distribution modeling, consisting of
occurrence and environmental data, from Halvorsen (2013). The study area
consists of 40 grid cells, with 8 row and 5 columns, in which 10 presences
occur.
}
\keyword{datasets}
