% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvfromev.R
\name{dvfromev}
\alias{dvfromev}
\title{produces dvs from a given ev by different transformations.}
\usage{
dvfromev(rv, ev, writedir, transformtype, allsplines)
}
\arguments{
\item{rv}{Response variable (vector) representing presence or background,
coded as: 1/NA.}

\item{ev}{Explanatory variable. May be continuous or categorical.}

\item{writedir}{Directory to which Maxent runs of spline transformations are
written}

\item{transformtype}{Set of transformation types to be used.}

\item{allsplines}{Logical. Keep all spline transformations.}
}
\value{
Dataframe with one column for each DV.
}
\description{
produces dvs from a given ev by different transformations.
}

