% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFOP.R
\name{plotFOP}
\alias{plotFOP}
\title{Plot Frequency of Observed Presence (FOP)}
\usage{
plotFOP(data, EV, intervals = 20, smoothwindow = 3, EVranging = FALSE)
}
\arguments{
\item{data}{Dataframe containing the response variable in the first column
and explanatory variables in subsequent columns. The response variable
should represent presence/background data, coded as: 1/NA. See Details for
information regarding presence/absence data.}

\item{EV}{Name or column number of the explanatory variable for which to
calculate FOP.}

\item{intervals}{Number of intervals into which the continuous EV is divided.
Irrelevant for categorical EVs.}

\item{smoothwindow}{Width of the smoothing window. Represents the number of
intervals included in an exponentially weighted moving average. Should be
odd, otherwise the window will be uncentered. Irrelevant for categorical
EVs.}

\item{EVranging}{if \code{TRUE}, will range the EV scale to [0,1]. This is
equivalent to plotting FOP over the linear transformation produced by
deriveVars. Irrelevant for categorical EVs.}
}
\value{
In addition to the plotted output, a list is returned containing 1)
  the EV value at which FOP is highest (\code{EVoptimum}) and 2) a data frame
  with the plotted data (\code{FOPdata}).
}
\description{
\code{plotFOP} produces a Frequency of Observed Presence (FOP) plot for a
given explanatory variable. For continuous variables, the exponetially
weighted moving average of the FOP values is added. \code{plotFOP} also
returns a list containing the optimum EV value, and a data frame with the
plotted data (for customizable plotting).
}
\details{
The EVoptimum that is retuned is based on the smoothed data, unless a maximum
exists at the extremes of the EV (outside the smoothing window). Note that if
the response variable represents presence/absence data, the result is an
empirical frequency of presence curve, rather than a observed frequency of
presence curve (see Stoea et al. [in press], Sommerfeltia).

\code{DESCRIPTION Imports}: dplyr, Hmisc, scales
}

