% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResp.R
\name{plotResp}
\alias{plotResp}
\title{Plot model response across a given explanatory variable.}
\usage{
plotResp(rv, ev, evdata, dvdata, writedir = NULL, jarpath = NULL)
}
\arguments{
\item{rv}{Response variable vector used to train the model. The RV should
represent presence/background data, coded as: 1/NA.}

\item{ev}{Name or list index of the explanatory variable in \code{dvdata} for
which the response curve is to be generated. Interaction terms not allowed.}

\item{evdata}{Explanatory variables used to train the model, in their
original (untransformed) form. Data frame.}

\item{dvdata}{Derived variables used to train the model. Named list of data
frames, with each data frame containing 1 or more DVs for a given EV. E.g.
output [[1]] of \code{selectEV}.}

\item{writedir}{The directory to which Maxent files will be written during
subset selection of DVs. Defaults to the working directory.}

\item{jarpath}{The pathway to the maxent.jar executable jar file. If
unspecified, the function looks for the file in the writedir.}
}
\value{
In addition to the graphical output, a data frame containing the
  plotted data is returned.
}
\description{
\code{plotResp} plots  the response of a given Maxent model over any of the
included explanatory variables (EVs) in that model. For categorical
variables, a box plot is returned rather than a curve. \code{plotResp} also
returns a data frame containing the plotted data (for customizable graphics).
}
\details{
The response curves generated in this function are single-effect response
curves, presenting the response of a model containing the explanatory
variable of interest only (cf. marginal-effect response curves).

The \code{ev} specified in \code{dvdata} must not be an interaction term.
}

