% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResp.R
\name{plotResp}
\alias{plotResp}
\title{Plot single-effect model response.}
\usage{
plotResp(data, dvdata, EV, dir = NULL, logscale = FALSE, ...)
}
\arguments{
\item{data}{Data frame containing the response variable in the first column
and explanatory variables in subsequent columns. The response variable
should represent presence/background data, coded as: 1/NA.}

\item{dvdata}{List of derived variables used to train the model, with each
list item a data frame containing 1 or more DVs for a given EV. E.g. output
[[1]] of \code{selectEV}.}

\item{EV}{Name or list index of the explanatory variable in \code{dvdata} for
which the response curve is to be generated. Interaction terms not allowed.}

\item{dir}{Directory to which files will be written. Defaults to the working
directory.}

\item{logscale}{Logical. Plot the common logarithm of PRO rather than PRO
itself.}

\item{...}{Arguments to be passed to \code{plot} to control the appearance of
the plot. For example: \itemize{ \item \code{cex} for size of points \item
\code{col} for color \item \code{pch} for type }}
}
\value{
In addition to the graphical output, the plotted data is returned. In
  the case of a continuous EV, the plotted data is contained in a list of 2:
  \enumerate{ \item Model response across individual data points
  (\code{respPts}). Columns in this data frame represent the following: EV
  value ("EV"), Probability Ration Output of the model ("PRO"), and
  corresponding EV interval ("int"). \item Model response across intervals of
  the EV (\code{respLine}). Columns in this data frame represent the
  following: EV interval ("int"), number of points in the interval ("n"),
  mean EV value of the points in the interval ("intEV"), mean Probability Ratio Output value of the
  points in the interval ("intPRO"), and exponentially weighted moving average
  of intPRO ("smoothPRO").}
}
\description{
\code{plotResp} plots the single-effect response of a given Maxent model over
any of the included explanatory variables (EVs) in that model. For
categorical variables, a bar plot is returned rather than a scatter plot.
\code{plotResp} also returns a data frame containing the plotted data (for
customizable graphics). Single-effect response curves present the response of
a model containing the explanatory variable of interest only (cf.
marginal-effect response curves; \code{\link{plotResp2}}).
}
\details{
The plot contains points, representing the model response across individual
data points, as well as a line, representing an exponentially weighted moving
average of the model response over intervals of the EV.

The \code{EV} specified in \code{dvdata} must not be an interaction term.
}

