% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResp2.R
\name{plotResp2}
\alias{plotResp2}
\title{Plot marginal-effect model response.}
\usage{
plotResp2(data, transformations, model, EV, logscale = FALSE, ...)
}
\arguments{
\item{data}{Data frame containing the observations used to train the
\code{model}, with the response variable in the first column and
explanatory variables in subsequent columns. The response variable should
represent presence/background data, coded as: 1/NA. See
\code{\link{readData}}.}

\item{transformations}{Transformation functions used to create the derived
variables in the model. I.e. the 'transformations' returned by
\code{\link{deriveVars}}. Equivalently, the full file pathway of the
'transformations.Rdata' file saved as a result of \code{\link{deriveVars}}.}

\item{model}{The model for which the response is to be plotted, represented
by an object of class 'glm'. This may be the object returned by
\code{\link{chooseModel}}, or the 'selectedmodel' returned by
\code{\link{selectEV}}.}

\item{EV}{Name of the explanatory variable or its column index in \code{data}
for which the response curve is to be plotted. Interaction terms not
allowed.}

\item{logscale}{Logical. Plot the common logarithm of PRO rather than PRO
itself.}

\item{...}{Arguments to be passed to \code{plot} or \code{barplot} to control
the appearance of the plot. For example: \itemize{ \item \code{lwd} for
line width \item \code{cex.main} for size of plot title \item \code{space}
for space between bars }}
}
\description{
\code{plotResp2} plots the marginal-effect response of a given model over any
of the explanatory variables (EVs) included in that model. For categorical
variables, a bar plot is returned rather than a scatter plot. Marginal-effect
response curves present the response of the model when all other explanatory
variables are held constant at their mean values (cf. single-effect response
curves; \code{\link{plotResp}}).
}
\examples{

}
