% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectModel.R
\name{projectModel}
\alias{projectModel}
\title{Project model to data.}
\usage{
projectModel(data, transformation, model, clamping = FALSE, rescale = FALSE)
}
\arguments{
\item{data}{Data frame of all the explanatory variables (EVs) included in the
model, with column names matching EV names. See \code{\link{readData}}.}

\item{transformation}{Full pathway of the 'transformations.Rdata' file
containing the transformations used to build the model. This file is saved
as a result of the \code{\link{deriveVars}} function. Equivalently, the
second item in the list returned by \code{\link{deriveVars}} can be used
directly.}

\item{model}{Full pathway of the '.lambdas' file of the model in question.
This file is saved as a result of \code{\link{selectEV}}.}

\item{clamping}{Logical. Do clamping \emph{sensu} Phillips et al. (2006).
Default is \code{FALSE}.}

\item{rescale}{Logical. Linearly rescale the probability ratio output (PRO)
to the range of values in \code{data}? This has implications for the
interpretation of model output with respect to reference value PRO = 1. See
details.}
}
\value{
List of 2: \enumerate{ \item A data frame with the model output in
  column 1 and the corresponding explanatory data in subsequent columns.
  \item A data frame showing the range of \code{data} compared to the
  training data, on a 0-1 scale.}
}
\description{
\code{projectModel} Calculates the probability ratio output (PRO) of a given
model for any points where values of the explanatory variables in the model
are known.The transformations performed on the explanatory variables to build
the model must be specified.
}
\details{
Missing data (NA) for a continuous variable will result in NA output for that
point. Missing data for a categorical variable is treated as belonging to
none of the categories.

When \code{rescale = FALSE}  the scale of the probability ratio output (PRO)
returned by this function is dependent on the data used to train the model.
As a result, values of PRO > 1 can be interpreted as higher relative
probability of presence than an average site in the \emph{training} data
(Halvorsen, 2013, Halvorsen et al., 2015). When \code{rescale = TRUE}, the
output is linearly rescaled so that values of PRO > 1 can be interpreted as
higher relative probability of presence than an average site in the
\emph{projection} data.
}
\examples{
\dontrun{
modeloutput <- projectModel(newdat,
   transformation = "D:/path/to/modeling/directory/deriveVars/transformations.Rdata",
   model = "D:/path/to/modeling/directory/selectEV/round/model/1.lambdas")
}

grasslandPrediction <- projectModel(grasslandPO,
   transformation = system.file("extdata", "transformations.Rdata", package = "maxentmodelselectr"),
   model = system.file("extdata", "1.lambdas", package = "maxentmodelselectr"))
head(grasslandPrediction$output)
grasslandPrediction$ranges

library(raster)
contfiles <- list.files(system.file("extdata", "EV_continuous", package = "maxentmodelselectr"), full.names = TRUE)
catfiles <- list.files(system.file("extdata", "EV_categorical", package = "maxentmodelselectr"), full.names = TRUE)
stack <- raster::stack(c(contfiles, catfiles))
stackpts <- rasterToPoints(stack)
spatialPrediction <- projectModel(stackpts,
   transformation = system.file("extdata", "transformations.Rdata", package = "maxentmodelselectr"),
   model = system.file("extdata", "1.lambdas", package = "maxentmodelselectr"))
Predictionraster <- raster(stack, layer=0)
Predictionraster <- rasterize(spatialPrediction$output[, c("x", "y")], Predictionraster, field = spatialPrediction$output$PRO)
plot(Predictionraster, colNA="black")

}
\references{
Halvorsen, R. (2013) A strict maximum likelihood explanation of
  MaxEnt, and some implications for distribution modelling. Sommerfeltia, 36,
  1-132.

Halvorsen, R., Mazzoni, S., Bryn, A. & Bakkestuen, V. (2015)
  Opportunities for improved distribution modelling practice via a strict
  maximum likelihood interpretation of MaxEnt. Ecography, 38, 172-183.

Phillips, S.J., Anderson, R.P. & Schapire, R.E. (2006) Maximum
  entropy modeling of species geographic distributions. Ecological Modelling,
  190, 231-259.
}

