% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectModel.R
\name{projectModel}
\alias{projectModel}
\title{Project model to data.}
\usage{
projectModel(newdata, transf, model, clamping = FALSE)
}
\arguments{
\item{newdata}{Data frame containing data for all the explanatory variables
included in the model.}

\item{transf}{Pathway to the .Rdata file containing the named parameterized
transformations used in the model. This file is saved as a result of the
\code{deriveVars} function.}

\item{model}{Pathway to the .lambdas file of the model in question. This file
is saved as a result of the \code{selectEV} function.}

\item{clamping}{logical. Do clamping. Default is \code{FALSE}.}
}
\value{
1) a data frame with the model output and the corresponding
  explanatory data. 2) a data frame showing the range of the data compared to
  the training data, on a 0-1 scale.
}
\description{
\code{projectModel} Calculates the probability ratio output (PRO) of a given
model for any points where values of the explanatory variables in the model
are known.The transformations performed on the explanatory variables to build
the model must be specified.
}

