% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectModel.R
\name{projectModel}
\alias{projectModel}
\title{Project model to data.}
\usage{
projectModel(data, transformation, model, clamping = FALSE)
}
\arguments{
\item{data}{Data frame containing data for all the explanatory variables
(EVs) included in the model, with column names matching EV names.}

\item{transformation}{Pathway to the 'transformations.Rdata' file containing
the transformations used to build the model. This file is saved as a result
of the \code{\link{deriveVars}} function. Equivalently, the second item in
the list returned by \code{\link{deriveVars}} can be used directly.}

\item{model}{Pathway to the '.lambdas' file of the model in question. This
file is saved as a result of \code{\link{selectEV}}.}

\item{clamping}{Logical. Do clamping \emph{sensu} maxent.jar. Default is
\code{FALSE}.}
}
\value{
List of 2: \enumerate{ \item A data frame with the model output in
  column 1 and the corresponding explanatory data in subsequent columns.
  \item A data frame showing the range of \code{data} compared to the
  training data, on a 0-1 scale.}
}
\description{
\code{projectModel} Calculates the probability ratio output (PRO) of a given
model for any points where values of the explanatory variables in the model
are known.The transformations performed on the explanatory variables to build
the model must be specified.
}
\details{
Missing data (NA) for a continuous variable will result in NA output for that
point. Missing data for a categorical variable is treated as belonging to
none of the categories.
}

