% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readData}
\alias{readData}
\title{Read in data from files.}
\usage{
readData(presence, contEV, catEV, maxbkg = 10000)
}
\arguments{
\item{presence}{Full pathway of the '.csv' file of occurrence data. The first
column codes presence of the model target, while the second and third
columns should contain X and Y coordinates cooresponding to the ASCII
raster coordinate system. The first row is read as a header column.}

\item{contEV}{Pathway to a directory containing continuous environmental
variables in '.asc' file format.}

\item{catEV}{Pathway to a directory containing categorical environmental
variables in '.asc' file format.}

\item{maxbkg}{Integer. Maximum number of grid cells selected as unknown
background points for the response varaible. Default is 10,000.}
}
\value{
Data frame with the Response Variable (RV) in the first column, and
  Explanatory Variables (EVs) in subsequent columns.
}
\description{
\code{readData} reads in occurence data in CSV format and environmental data
in ASCII raster format and produces a data object which can be used as the
starting point for many analyses in this package. This function is intended
to make reading in data easy for users familiar with the maxent.jar program.
It emphasized that important considerations for data preparation (e.g.
cleaning, sampling bias removal, etc.) are not treated in this package and
must be dealt with separately!
}
\details{
Background points are randomly selected from the full extent of the grid
files, and may include presence locations. Only cells which contain data for
all variables are selected.
}

