% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectDV.R
\name{selectDV}
\alias{selectDV}
\title{Select parsimonious sets of derived variables.}
\usage{
selectDV(dv, rv, writedir = NULL, jarpath = NULL)
}
\arguments{
\item{dv}{List of data frames, with each data frame containing DVs for a
given EV.}

\item{rv}{Response variable vector. The RV should represent
presence/background data, coded as: 1/NA.}

\item{writedir}{The directory to which Maxent files will be written during
subset selection of DVs. Defaults to the working directory.}

\item{jarpath}{The pathway to the maxent.jar executable jar file. If
unspecified, the function looks for the file in the writedir.}
}
\value{
List of data frames, with each data frame containing \emph{selected}
  DVs for a given EV. Dimensions: [[EV]] [N, <=DV].
}
\description{
For each explanatory variable, \code{selectDV} selects the set of derived
variables which best explains variation in a given response variable.
}
\details{
If the derived variables were created using \code{\link{deriveVars}}, the
same response variable should be used in \code{selectDV}, as the deviation
and spline transformations produced by \code{deriveVars} are RV-specific.
}
\references{
Halvorsen, R., Mazzoni, S., Bryn, A., & Bakkestuen, V. (2015).
  Opportunities for improved distribution modelling practice via a strict
  maximum likelihood interpretation of MaxEnt. Ecography, 38(2), 172-183.
}

