% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectEV.R
\name{selectEV}
\alias{selectEV}
\title{Select parsimonious sets of explanatory variables.}
\usage{
selectEV(data, dvdata, alpha = 0.01, interaction = TRUE, dir = NULL,
  jar = NULL, trainmax = NULL)
}
\arguments{
\item{data}{Data frame containing the response variable in the first column
and explanatory variables in subsequent columns. The response variable
should represent presence/background data, coded as: 1/NA.}

\item{dvdata}{List of data frames, with each data frame containing a
parsimonious group of derived variables for a given explanatory variable
(e.g. the first item in the list returned by \code{\link{selectDVforEV}}:
\code{selectDVforEV(...)[[1]]})}

\item{alpha}{Alpha-level used in F-test comparison of models. Default is
0.01.}

\item{interaction}{Logical. Allows interaction terms between pairs of EVs.
Default is \code{TRUE}.}

\item{dir}{Directory to which files will be written during subset
selection of explanatory variables. Defaults to the working directory.}

\item{jar}{Pathway to the 'maxent.jar' executable jar file. If
unspecified, the function looks for the file in \code{dir}.}

\item{trainmax}{Integer. Maximum number of uninformed background points to be
used to train the models. May be used to reduce computation time for data
sets with very large numbers of points. Default is no maximum.}
}
\value{
List of 2: \enumerate{ \item A list of data frames, with one
  data frame for each \emph{selected} EV. \item A data frame
  showing the trail of forward selection of individual EVs (and interaction
  terms if necessary).}
}
\description{
\code{selectEV} selects the parsimonious set of explanatory variables (EVs)
which best explains variation in a given response variable (RV). Each EV can
be represented by 1 or more derived variables (see \code{\link{deriveVars}}).
The function uses a process of forward selection based on comparison of
nested models by the F-test, where the F-statistic is calculated using
equation 59 in Halvorsen (2013). See Halvorsen et al. (2015) for an
explanation of the forward selection procedure.
}
\details{
When \code{interaction = TRUE}, the forward selection procedure selects a
parsimonious group of individual EVs first, and then tests interactions
between EVs included in the model afterwards. Therefore, interactions are
only explored between terms which are individually explain a significant
amount of variation. When \code{interaction = FALSE}, interactions are not
considered.

Variables should be uniquely named, and the names should not contain spaces
or colons.
}
\references{
Halvorsen, R. (2013). A strict maximum likelihood explanation of
  MaxEnt, and some implications for distribution modelling. Sommerfeltia, 36,
  1-132.

Halvorsen, R., Mazzoni, S., Bryn, A., & Bakkestuen, V. (2015).
  Opportunities for improved distribution modelling practice via a strict
  maximum likelihood interpretation of MaxEnt. Ecography, 38(2), 172-183.
}

