% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectEV.R
\name{selectEV}
\alias{selectEV}
\title{Select parsimonious sets of explanatory variables.}
\usage{
selectEV(rv, ev, alpha = 0.01, interaction = TRUE, writedir = NULL,
  jarpath = NULL)
}
\arguments{
\item{rv}{Response variable vector. The RV should represent
presence/background data, coded as: 1/NA.}

\item{ev}{Named list of data frames, with each data frame containing 1 or
more DVs for a given EV. E.g. output [[1]] of \code{selectDV}.}

\item{alpha}{Alpha-level used in F-test comparison of models. Default is
0.01.}

\item{interaction}{Logical. Allows interaction terms between pairs of EVs.
Default is \code{TRUE}.}

\item{writedir}{The directory to which Maxent files will be written during
subset selection of DVs. Defaults to the working directory.}

\item{jarpath}{The pathway to the maxent.jar executable jar file. If
unspecified, the function looks for the file in the writedir.}
}
\value{
List of length two. The first item is a list of data frames, with one
  data frame for each \emph{selected} EV. The second item is also a list of
  data frames, where the first shows the trail of forward selection of
  individual EVs, while the second shows the trail of forward selection of
  interaction terms (if necessary).
}
\description{
\code{selectEV} selects the parsimonious set of explanatory variables (EVs)
which best explains variation in a given response variable (RV). Each EV can
be represented by 1 or more derived variables (see \code{\link{deriveVars}}).
The function uses a process of forward selection based on comparison of
nested models by the F-test, where the F-statistic is calculated using
equation 59 in Halvorsen (2013). See Halvorsen et al. (2015) for an
explanation of the forward selection procedure.
}
\details{
When \code{interaction = TRUE}, the forward selection procedure selects a
parsimonious group of individual EVs first, and then tests interactions
between EVs included in the model afterwards. Therefore, interactions are
only explored between terms which are individually explain a significant
amount of variation. When \code{interaction = FALSE}, interactions are not
considered.

Each item in the EV list must be uniquely named, and the names must not
contain spaces.
}
\references{
Halvorsen, R. (2013). A strict maximum likelihood explanation of
  MaxEnt, and some implications for distribution modelling. Sommerfeltia, 36,
  1-132.

Halvorsen, R., Mazzoni, S., Bryn, A., & Bakkestuen, V. (2015).
  Opportunities for improved distribution modelling practice via a strict
  maximum likelihood interpretation of MaxEnt. Ecography, 38(2), 172-183.
}

