% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testAUC.R
\name{testAUC}
\alias{testAUC}
\title{Calculate model AUC with test data.}
\usage{
testAUC(data, transformation, model)
}
\arguments{
\item{data}{Data frame containing test data in the first column and
corresponding explanatory variables in subsequent columns. The test data
should be coded as: 1/0/NA, representing presence, absence, and unknown.}

\item{transformation}{Pathway to the 'transformation.Rdata' file containing
the transformations used to build the model. This file is saved as a result
of the \code{\link{deriveVars}} function. Equivalently, the second item in
the list returned by \code{\link{deriveVars}} can be used directly.}

\item{model}{Pathway to the '.lambdas' file of the model in question. This
file is saved as a result of \code{\link{selectEV}}.}
}
\value{
In addition to returning the testAUC value, graphical output showing
  the corresponding ROC plot is produced.
}
\description{
For a given model, \code{testAUC} calculates the Area Under the Curve (AUC)
of the Receiver Operating Characteristic (ROC) as a threshold-independent
measure of binary classification performance. This function is intended to be
used with occurence data that is independent from the data used to train the
model, to obtain an unbiased measure of model performance.
}
\details{
\code{DESCRIPTION Imports}: PresenceAbsence
}

