% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toydata_selevs}
\alias{toydata_selevs}
\title{Selected explanatory variables accompanied by selection trails, from toy
data.}
\format{List with 3 elements: \enumerate{ \item A list of 3, with the
  response variable followed by data frames, represent selected explanatory
  variables. \item A trail of forward model selection used to select
  explanatory variables and interaction terms. \item The selected model }}
\source{
Produced from \code{\link{toydata_seldvs}} using
  \code{\link{selectEV}}.
}
\usage{
toydata_selevs
}
\description{
Selected explanatory variables and tables showing forward model selection of
explanatory variables for distribution modeling of a small, synthetic data
set used in Halvorsen (2013). Each individual explanatory variable is
represented by a group of derived variables.
}
\references{
Halvorsen, R. (2013) A strict maximum likelihood explanation of
  MaxEnt, and some implications for distribution modelling. Sommerfeltia, 36,
  1-132.
}
\keyword{datasets}
